[Fixed] Compiler Warnings
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@877 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/object/ObjectControllerMessage.cpp
@@ -2411,7 +2411,7 @@ void ObjectControllerMessage::parseResourceEmptyHopper(Player* player, Message* 
 
 	bool makeNewResource = true;
 
-	quantity = inso->removeHopperItem(rId, quantity);
+	quantity = (uint32)inso->removeHopperItem(rId, quantity);
 	if(quantity >= 1)
 	{
 		ResourceContainer* newRcno = new ResourceContainer(player->getNewItemID());
